# src/main/java/io/swagger/petstore/model/ApiResponse.java
package io.swagger.petstore.model;

import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;

/**
 * Generic API response wrapper. The spec names the root XML element "##default",
 * but the reference server emits <ApiResponse> for XML, so we match that behavior.
 */
@JacksonXmlRootElement(localName = "ApiResponse")
public class ApiResponse {
    private Integer code;
    private String type;
    private String message;

    public ApiResponse() {}

    public ApiResponse(Integer code, String type, String message) {
        this.code = code;
        this.type = type;
        this.message = message;
    }

    public Integer getCode() { return code; }
    public void setCode(Integer code) { this.code = code; }
    public String getType() { return type; }
    public void setType(String type) { this.type = type; }
    public String getMessage() { return message; }
    public void setMessage(String message) { this.message = message; }
}